Added Full LG Audio Configuration
You were right, it was using stock lineage audio settings. No wonder it was so quiet. I also upped the API level on gestures and doze. Basically, any app that had the wrong API level. No more crashes I hope.

Removed the Wi-Fi aware as it was not properly configured.